move parallelizer to ingestion pkg
