+++ pkg/ingestion/paralellizer.go
@@ -1,4 +1,4 @@
-package remotewrite
+package ingestion
 
 import (
 	"context"
@@ -7,24 +7,21 @@ import (
 	"sync"
 
 	"github.com/sirupsen/logrus"
-
-	"github.com/pyroscope-io/pyroscope/pkg/ingestion"
 )
 
-// TODO(eh-am): move this to somehwere else?
 type Parallelizer struct {
 	log       *logrus.Logger
-	ingesters []ingestion.Ingester
+	ingesters []Ingester
 }
 
-func NewParallelizer(log *logrus.Logger, ingesters ...ingestion.Ingester) *Parallelizer {
+func NewParallelizer(log *logrus.Logger, ingesters ...Ingester) *Parallelizer {
 	return &Parallelizer{
 		log:       log,
 		ingesters: ingesters,
 	}
 }
 
-func (p *Parallelizer) Ingest(ctx context.Context, in *ingestion.IngestInput) error {
+func (p *Parallelizer) Ingest(ctx context.Context, in *IngestInput) error {
 	var wg sync.WaitGroup
 	wg.Add(len(p.ingesters))
 
@@ -32,13 +29,15 @@ func (p *Parallelizer) Ingest(ctx context.Context, in *ingestion.IngestInput) er
 	for _, putter := range p.ingesters {
 		// https://golang.org/doc/faq#closures_and_goroutines
 		putter := putter
-		go func(in *ingestion.IngestInput) {
+		in := in
+
+		go func() {
 			defer wg.Done()
 			err := p.safeIngest(ctx, in, putter)
 			if err != nil {
 				p.log.Error("Failed to parallelize put: ", err)
 			}
-		}(in)
+		}()
 	}
 
 	wg.Wait()
@@ -47,7 +46,7 @@ func (p *Parallelizer) Ingest(ctx context.Context, in *ingestion.IngestInput) er
 }
 
 // This is required since ingester.Ingest may panic
-func (*Parallelizer) safeIngest(ctx context.Context, input *ingestion.IngestInput, ingester ingestion.Ingester) (err error) {
+func (*Parallelizer) safeIngest(ctx context.Context, input *IngestInput, ingester Ingester) (err error) {
 	defer func() {
 		if r := recover(); r != nil {
 			err = fmt.Errorf("panic recovered: %v; %v", r, string(debug.Stack()))
